+++ changelog.md
@@ -5,3 +5,8 @@
 - Suggested nodes to be updated are based on the `Value Node` provided by [pySHACL](https://github.com/RDFLib/pySHACL)
 - Include paths of the RDF graphs to be validated and SHACL files, as well as the datetime of execution
 - Text report is divided into sections based on the nodes suggested for update to enhance readability
+
+## v0.3.3
+
+- Update visualizations: supports `txt` and `html` output formats
+- Replace `PyGraphviz` with `NetworkX` and `PyVis` packages